Fix six of ten correctness defects: fade-window UB, MultiSync ordering, truncated identity, and gap accounting
Adds a mutex around BrightnessEngine (finding 1 wiring lands with the
adapters next), fade-window validation on adoptState, range-checked
numeric decoding, an inverted-persisted-window fix that settles darker,
ObservationSink acceptance semantics, MultiSync ordering by
(stateChangedAtMillis, instanceId), truncated-identity detection, and
the lost-wakeup fix on the worker's condition variable.

+++ native/include/showmesh/fading_value.h
@@ -50,8 +50,12 @@ class FadingValue {
         if (now >= endMillis_) {
             return target_;
         }
-        const double elapsed = static_cast<double>(now - startMillis_);
-        const double span = static_cast<double>(endMillis_ - startMillis_);
+        // Subtracting as int64 first can overflow (undefined behavior) for
+        // an adopted window with extreme endpoints; subtracting as double
+        // cannot, regardless of what validation ran before this value was
+        // restored.
+        const double elapsed = static_cast<double>(now) - static_cast<double>(startMillis_);
+        const double span = static_cast<double>(endMillis_) - static_cast<double>(startMillis_);
         return start_ + (target_ - start_) * (elapsed / span);
     }
